[department-of-veterans-affairs/orchid#68] Refactors formConfig into separate pages
