+++ src/applications/ask-a-question/config/pages/index.js
@@ -1,0 +1,4 @@
+import contactInformationPage from './contactInformationPage';
+import inquiryPage from './inquiryPage';
+
+export { contactInformationPage, inquiryPage };
